Transcriber/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128
torch
torchvision
torchaudio


openai-whisper
pydub==0.25.1
audioop-lts 
